+++ Form/Types/ManagedUploadType.php
@@ -122,6 +122,7 @@ class ManagedUploadType extends AbstractType
                 ]);
             }
         });
+
     }
 
     /**